+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -4,8 +4,8 @@ import nextstep.subway.applicaion.dto.*;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
 import nextstep.subway.domain.Station;
-import nextstep.subway.exception.DuplicateLineException;
-import nextstep.subway.exception.LineNotFoundException;
+import nextstep.subway.exception.line.DuplicateLineException;
+import nextstep.subway.exception.line.LineNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
