+++ src/main/java/nextstep/subway/exception/ApplicationException.java
@@ -2,9 +2,6 @@ package nextstep.subway.exception;
 
 public class ApplicationException extends RuntimeException {
 
-    public ApplicationException() {
-    }
-
     public ApplicationException(String message) {
         super(message);
     }
