+++ src/main/java/nextstep/subway/exception/section/AlreadyRegisteredStationInLineException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.exception.section;
 
 public class AlreadyRegisteredStationInLineException extends RuntimeException {
