+++ src/main/java/nextstep/subway/exception/section/DownStationNotMatchException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.exception.section;
 
 public class DownStationNotMatchException extends RuntimeException {
