+++ src/main/java/nextstep/subway/exception/station/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.exception.station;
 
 public class StationNotFoundException extends RuntimeException {
